finished diploma project

+++ app/src/main/java/com/example/qrwithjetpack/presentation/feature/photoPreview/PhotoPreviewScreen.kt
@@ -1,5 +1,9 @@
 package com.example.qrwithjetpack.presentation.feature.photoPreview
 
+import android.annotation.SuppressLint
+import android.os.Build
+import android.util.Log
+import androidx.annotation.RequiresApi
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -11,11 +15,20 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.sizeIn
 import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Button
+import androidx.compose.material3.SnackbarHost
+import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.platform.LocalContext
@@ -24,8 +37,11 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
 import com.example.qrwithjetpack.presentation.feature.cameraPreview.CameraPreviewViewModel
 import com.example.qrwithjetpack.util.Util
+import kotlinx.coroutines.launch
 
 
+@RequiresApi(Build.VERSION_CODES.S)
+@SuppressLint("CoroutineCreationDuringComposition")
 @Composable
 fun PhotoPreviewScreen(
     navController: NavController,
@@ -34,44 +50,85 @@ fun PhotoPreviewScreen(
 ) {
     val capturedPhoto: ImageBitmap = lastCapturedPhoto!!
     val context = LocalContext.current
+    val imagesWithNoFace = cameraPreviewViewModel.imagesWithNoFace.collectAsState(initial = null).value
+    val imageProcessionEnded = cameraPreviewViewModel.imageAnalyzer.imageProcessionEnded.collectAsState(
+        initial = false
+    ).value
+    val imageHasFace = cameraPreviewViewModel.imageAnalyzer.imageHasFace.collectAsState(
+        initial = false
+    ).value
+    val faceFeatures = cameraPreviewViewModel.imageAnalyzer.faceFeatures.collectAsState(
+        initial = floatArrayOf()
+    ).value
+
+    val scope = rememberCoroutineScope()
+    val snackbarHostState = remember { mutableStateOf(SnackbarHostState()) }
 
     Column(
         modifier = Modifier.fillMaxSize(),
         verticalArrangement = Arrangement.SpaceBetween
     ) {
+        SnackbarHost(snackbarHostState.value)
         Box(
             modifier = Modifier
                 .fillMaxSize()
-                .weight(1f)
-                .border(20.dp, Color.Black) //
+                .weight(14f)
+                .border(10.dp, Color.Black) //
         ) {
-            Image(
-                bitmap = capturedPhoto,
-                contentDescription = "Last captured photo",
-                contentScale = androidx.compose.ui.layout.ContentScale.Crop
-            )
+
+            Box(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .sizeIn(20.dp, 30.dp)
+                    .border(10.dp, Color.Black) //
+            ) {
+                if (imageProcessionEnded && !imageHasFace) {
+                    scope.launch {
+                        snackbarHostState.value.showSnackbar("Суретте бет жоқ! Қайттан суретке түсіріңіз.")
+                    }
+                } else if (imageProcessionEnded && imageHasFace) {
+                    Log.i("Detect Face Features",
+                        (System.currentTimeMillis() - cameraPreviewViewModel.timeToDetectFaceFeatures).toString() + " мс"
+                    )
+                    scope.launch {
+                        snackbarHostState.value.showSnackbar("Суретте бет бар! QR-код жасауға рұқсат.")
+                    }
+                }
+            }
+            Row(modifier = Modifier
+                .padding(2.dp)
+            ) {
+                Image(
+                    bitmap = capturedPhoto,
+                    contentDescription = "Соңғы түсірілген сурет",
+                    contentScale = androidx.compose.ui.layout.ContentScale.FillWidth,
+                    modifier = Modifier.clip(RoundedCornerShape(15.dp)),
+
+                )
+            }
         }
         Row(
             modifier = Modifier
                 .fillMaxWidth()
-//
                 .background(Color.Black)
-                .padding(16.dp)
+                .padding(12.dp)
         ) {
-            Button(
-                onClick = {
-//                    capturePhoto(context, cameraController, imageCapture)
+            if (imageProcessionEnded && imageHasFace) {
+                Button(
+                    onClick = {
 
-                    val qrId = cameraPreviewViewModel.createQrFromPhoto(capturedPhoto)
-                    cameraPreviewViewModel.deleteFile(context)
+                        val qrId = cameraPreviewViewModel.createQrFromPhoto(capturedPhoto, faceFeatures)
 
-                    navController.navigate(Util.QRpreview_ROUTE + "/$qrId")
-                },
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .weight(1f)
-            ) {
-                Text(text = "QR-ға айналдыру")
+                        cameraPreviewViewModel.deleteFile(context)
+
+                        navController.navigate(Util.QRpreview_ROUTE + "/$qrId")
+                    },
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .weight(1f)
+                ) {
+                    Text(text = "QR жасау")
+                }
             }
             Spacer(modifier = Modifier.width(16.dp))
             Button(
@@ -88,11 +145,3 @@ fun PhotoPreviewScreen(
         }
     }
 }
-
-//@Preview
-//@Composable
-//fun Preview_Photo(
-////    navController: NavController
-//) {
-//    PhotoPreviewScreen()
-//}
